Refactor context and provider structure for improved organization
- Moved context providers for HeroSection and Models to a new hooks structure for better modularity.
- Updated imports across components to utilize the new hooks.
- Removed outdated context files to streamline the codebase.
- Adjusted ThemeProvider and RefreshContext to align with the new structure.
