+++ app/frontend/src/components/DeployModelStep.tsx
@@ -7,8 +7,8 @@ import { AnimatedDeployButton } from "./magicui/AnimatedDeployButton";
 import { useStepper } from "./ui/stepper";
 import { Weight } from "./SelectionSteps";
 import { StepperFormActions } from "./StepperFormActions";
-import { useModels } from "../providers/ModelsContext";
-import { useRefresh } from "../providers/RefreshContext";
+import { useModels } from "../hooks/useModels";
+import { useRefresh } from "../hooks/useRefresh";
 import { Cpu, Sliders, AlertTriangle, ExternalLink } from "lucide-react";
 import { checkCurrentlyDeployedModels } from "../api/modelsDeployedApis";
 import { Button } from "./ui/button";
