+++ app/frontend/src/components/aiPlaygroundHome/DeployedHome.tsx
@@ -11,7 +11,7 @@ import { Button } from "../ui/button";
 import { TypewriterEffectSmooth } from "../ui/typewriter-effect";
 // import { LineShadowText } from "../ui/line-shadow-text";
 // import { useState, useEffect } from "react";
-import { useHeroSection } from "../../providers/HeroSectionContext";
+import { useHeroSection } from "../../hooks/useHeroSection";
 
 export function DeployedHome({ onlyCards = false }: { onlyCards?: boolean } = {}) {
   const { showHero } = useHeroSection();
