+++ app/frontend/src/components/speechToText/appSidebar.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Search, Settings, Plus, MessageSquare, Moon, Sun } from "lucide-react";
 import { Button } from "../ui/button";
-import { useTheme } from "../../providers/ThemeProvider";
+import { useTheme } from "../../hooks/useTheme";
 
 import {
   Sidebar,
