+++ app/frontend/src/contexts/HeroSectionContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+export type HeroSectionContextType = {
+  showHero: boolean;
+  setShowHero: (val: boolean) => void;
+};
+
+export const HeroSectionContext = createContext<HeroSectionContextType | undefined>(undefined);
