+++ app/frontend/src/hooks/useModels.ts
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
+import { useContext } from "react";
+import { ModelsContext } from "../contexts/ModelsContext";
+
+export const useModels = () => {
+  const context = useContext(ModelsContext);
+  if (context === undefined) {
+    throw new Error("useModels must be used within a ModelsProvider");
+  }
+  return context;
+};
