+++ app/frontend/src/hooks/useRefresh.ts
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+import { useContext } from "react";
+import { RefreshContext, RefreshContextType } from "../contexts/RefreshContext";
+
+export const useRefresh = (): RefreshContextType => {
+  const context = useContext(RefreshContext);
+  if (context === undefined) {
+    throw new Error("useRefresh must be used within a RefreshProvider");
+  }
+  return context;
+};
