+++ app/frontend/src/hooks/useTheme.ts
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
+import { useContext } from "react";
+import { ThemeProviderContext, ThemeProviderState } from "../contexts/ThemeContext";
+
+export const useTheme = (): ThemeProviderState => {
+  const context = useContext(ThemeProviderContext);
+
+  if (context === undefined) throw new Error("useTheme must be used within a ThemeProvider");
+
+  return context;
+};
