+++ app/frontend/src/providers/ModelsContext.tsx
@@ -1,26 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
 
-import React, { createContext, useContext, useState, useCallback } from "react";
+import React, { useState, useCallback } from "react";
 import { fetchModels, fetchDeployedModelsInfo } from "../api/modelsDeployedApis";
-
-export interface Model {
-  id: string;
-  name: string;
-  image: string;
-  status: string;
-  health: string;
-  ports: string;
-}
-
-interface ModelsContextType {
-  models: Model[];
-  setModels: React.Dispatch<React.SetStateAction<Model[]>>;
-  refreshModels: () => Promise<void>;
-  hasDeployedModels: boolean;
-}
-
-const ModelsContext = createContext<ModelsContextType | undefined>(undefined);
+import { ModelsContext, type Model } from "../contexts/ModelsContext";
 
 export const ModelsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [models, setModels] = useState<Model[]>([]);
@@ -82,12 +65,4 @@ export const ModelsProvider: React.FC<{ children: React.ReactNode }> = ({ childr
       {children}
     </ModelsContext.Provider>
   );
-};
-
-export const useModels = () => {
-  const context = useContext(ModelsContext);
-  if (context === undefined) {
-    throw new Error("useModels must be used within a ModelsProvider");
-  }
-  return context;
 };
