+++ app/frontend/src/providers/RefreshContext.tsx
@@ -1,15 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
-import React, { createContext, useContext, useState } from "react";
-
-interface RefreshContextType {
-  refreshTrigger: number;
-  triggerRefresh: () => void;
-  triggerHardwareRefresh: () => Promise<void>; // New function for hardware cache refresh
-}
-
-const RefreshContext = createContext<RefreshContextType | undefined>(undefined);
+import React, { useState } from "react";
+import { RefreshContext } from "../contexts/RefreshContext";
 
 export const RefreshProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [refreshTrigger, setRefreshTrigger] = useState(0);
@@ -45,12 +38,4 @@ export const RefreshProvider: React.FC<{ children: React.ReactNode }> = ({ child
       {children}
     </RefreshContext.Provider>
   );
-};
-
-export const useRefresh = () => {
-  const context = useContext(RefreshContext);
-  if (context === undefined) {
-    throw new Error("useRefresh must be used within a RefreshProvider");
-  }
-  return context;
 };
